refactor(rate-service): rename base package to com.currencyconverter.rateservice
  details:
   - migrated all Java source and test files to the new directory structure mirroring the namespace
   - refactored the build configuration to use the shared currency-converter-parent
   - transitioned the web layer from blocking Spring Web to reactive Spring WebFlux
   - updated logging levels in environment-specific property files to reflect the new namespace

  impact:
   - aligns the rate service with the project's reactive architecture and naming standards
   - simplifies dependency management by leveraging the centralized parent POM
   - prepares the service for high-concurrency request processing in a non-blocking environment
