+++ rate-service/src/main/java/com/currencyconverter/rateservice/controller/StatusController.java
@@ -1,9 +1,11 @@
-package com.example.rateservice.controller;
+package com.currencyconverter.rateservice.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
 import java.util.Map;
 
 @RestController
@@ -12,8 +14,8 @@ public class StatusController {
   private static final Logger logger = LoggerFactory.getLogger(StatusController.class);
 
   @GetMapping("/status")
-  public Map<String, String> getStatus() {
+  public Mono<Map<String, String>> getStatus() {
     logger.info("Health check requested for rate-service");
-    return Map.of("status", "UP");
+    return Mono.just(Map.of("status", "UP"));
   }
 }
