+++ rate-service/src/main/java/com/currencyconverter/rateservice/exception/CurrencyNotSupportedException.java
@@ -1,4 +1,4 @@
-package com.example.rateservice.exception;
+package com.currencyconverter.rateservice.exception;
 
 public class CurrencyNotSupportedException extends RuntimeException {
   public CurrencyNotSupportedException(String message) {
