+++ rate-service/src/main/java/com/currencyconverter/rateservice/exception/ExchangeRateFetchException.java
@@ -1,4 +1,4 @@
-package com.example.rateservice.exception;
+package com.currencyconverter.rateservice.exception;
 
 public class ExchangeRateFetchException extends RuntimeException {
   public ExchangeRateFetchException(String message) {
