+++ rate-service/src/test/java/com/currencyconverter/rateservice/RateServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.rateservice;
+package com.currencyconverter.rateservice;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
