+++ rate-service/src/test/resources/application-test.properties
@@ -11,4 +11,4 @@ exchange-rate-api.connect-timeout=5000
 exchange-rate-api.read-timeout=5000
 
 # Logging
-logging.level.com.example.rateservice=DEBUG
+logging.level.com.currencyconverter.rateservice=DEBUG
